requirements.txt
redis
python-dotenv
pytest-asyncio
requests